Struct relay_wstring::WStr

#[repr(transparent)]pub struct WStr { /* fields omitted */ }

A UTF-16 str-like type with little-endian byte order.

This mostly behaves like str does for UTF-8 encoded bytes slices, but works with UTF-16LE encoded byte slices.

Examples

use relay_wstring::WStr;

let b = b"h\x00e\x00l\x00l\x00o\x00";
let s: &WStr = WStr::from_utf16le(b).unwrap();

let chars: Vec<char> = s.chars().collect();
assert_eq!(chars, vec!['h', 'e', 'l', 'l', 'o']);

assert_eq!(s.to_utf8(), "hello");

Implementations

impl WStr

pub fn from_utf16le(raw: &[u8]) -> Result<&Self, Utf16Error>

Creates a new &WStr from an existing UTF-16 little-endian encoded byte-slice.

If the byte-slice is not valid Utf16Error is returned.

pub fn from_utf16le_mut(raw: &mut [u8]) -> Result<&mut Self, Utf16Error>

Creates a new &mut WStr from an existing UTF-16 little-endian encoded byte-slice.

If the byte-slice is not valid Utf16Error is returned.

pub unsafe fn from_utf16le_unchecked(raw: &[u8]) -> &Self

Creates a new WStr from an existing UTF-16 little-endian encoded byte-slice.

Safety

You must guarantee that the buffer passed in is encoded correctly otherwise you will get undefined behavior.

pub unsafe fn from_utf16le_unchecked_mut(raw: &mut [u8]) -> &mut Self

Like WStr::from_utf16le_unchecked but return a mutable reference.

Safety

You must guarantee that the buffer passed in is encoded correctly otherwise you will get undefined behavior.

pub fn len(&self) -> usize

The length in bytes, not chars or graphemes.

pub fn is_empty(&self) -> bool

Returns true if the WStr::len is zero.

pub fn is_char_boundary(&self, index: usize) -> bool

Returns true if the index into the bytes is on a char boundary.

pub fn as_bytes(&self) -> &[u8]

Converts to a byte slice.

pub unsafe fn as_bytes_mut(&mut self) -> &mut [u8]

Converts to a mutable byte slice.

Safety

When mutating the bytes it must still be valid little-endian encoded UTF-16 otherwise you will get undefined behavior.

pub const fn as_ptr(&self) -> *const u8

Converts to a raw pointer to the byte slice.

pub fn as_mut_ptr(&mut self) -> *mut u8

Converts to a mutable raw pointer to the byte slice.

pub fn get<I>(&self, index: I) -> Option<&<I as SliceIndex<WStr>>::Output> where
I: SliceIndex<WStr>,

Returns a subslice of WStr.

The slice indices are on byte offsets of the underlying UTF-16LE encoded buffer, if the subslice is not on character boundaries or otherwise invalid this will return None.

pub fn get_mut<I>(
&mut self,
index: I
) -> Option<&mut <I as SliceIndex<WStr>>::Output> where
I: SliceIndex<WStr>,

Returns a mutable subslice of WStr.

The slice indices are on byte offsets of the underlying UTF-16LE encoded buffer, if the subslice is not on character boundaries or otherwise invalid this will return None.

pub unsafe fn get_unchecked<I>(
&self,
index: I
) -> &<I as SliceIndex<WStr>>::Output where
I: SliceIndex<WStr>,

Returns a subslice of WStr.

Safety

Like WStr::get but this results in undefined behavior if the sublice is not on character boundaries or otherwise invalid.

pub unsafe fn get_unchecked_mut<I>(
&mut self,
index: I
) -> &mut <I as SliceIndex<WStr>>::Output where
I: SliceIndex<WStr>,

Returns a mutable subslice of WStr.

Safety

Like WStr::get_mut but this results in undefined behavior if the subslice is not on character boundaries or otherwise invalid.

pub fn chars(&self) -> WStrChars<'_>

Returns an iterator of the chars of a string slice.

pub fn char_indices(&self) -> WStrCharIndices<'_>

Returns and iterator over the chars of a string slice and their positions.

pub fn to_utf8(&self) -> String

Returns the WStr as a new owned String.

pub fn is_ascii(&self) -> bool

Returns true if all characters in the string are ASCII.