Trait relay_wstring::SliceIndex

pub trait SliceIndex<T>: SealedSliceIndex where
    T: ?Sized, {
    type Output: ?Sized;
    fn get(self, slice: &T) -> Option<&Self::Output>;

    fn get_mut(self, slice: &mut T) -> Option<&mut Self::Output>;

    unsafe fn get_unchecked(self, slice: &T) -> &Self::Output;

    unsafe fn get_unchecked_mut(self, slice: &mut T) -> &mut Self::Output;

    fn index(self, slice: &T) -> &Self::Output;

    fn index_mut(self, slice: &mut T) -> &mut Self::Output;
}

Our own version of std::slice::SliceIndex.

Since this is a sealed trait, we need to re-define this trait. This trait itself is sealed as well.

Associated Types

type Output: ?Sized

The result of slicing, another slice of the same type as you started with normally.

Loading content...
Required methods

fn get(self, slice: &T) -> Option<&Self::Output>

Returns a shared reference to the output at this location, if in bounds.

fn get_mut(self, slice: &mut T) -> Option<&mut Self::Output>

Returns a mutable reference to the output at this location, if in bounds.

unsafe fn get_unchecked(self, slice: &T) -> &Self::Output

Like Self::get but without bounds checking.

Safety

You must guarantee the resulting slice is valid UTF-16LE, otherwise you will get undefined behavour.

unsafe fn get_unchecked_mut(self, slice: &mut T) -> &mut Self::Output

Like Self::get_mut but without bounds checking.

Safety

You must guarantee the resulting slice is valid UTF-16LE, otherwise you will get undefined behavour.

fn index(self, slice: &T) -> &Self::Output

Returns a shared reference to the output at this location, panicking if out of bounds.

fn index_mut(self, slice: &mut T) -> &mut Self::Output

Returns a mutable reference to the output at this location, panicking if out of bounds.

Loading content...